#284 refactor naming
