+++ server/balances_history_cacher/Dockerfile
