+++ server/balances_history_cacher/balance_caching_app/apps.py
@@ -1,10 +1,10 @@
 from django.apps import AppConfig
 
 
-class BalanceCashingAppConfig(AppConfig):
+class BalanceCachingAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'balance_cashing_app'
+    name = 'balance_caching_app'
 
     def ready(self):
-        from balance_cashing_app import scheduler
+        from balance_caching_app import scheduler
         scheduler.start()
