+++ server/balances_history_cacher/balance_caching_app/scheduler.py
@@ -5,21 +5,20 @@ import requests
 from apscheduler.schedulers.background import BackgroundScheduler
 from django.utils import timezone
 
-from balance_cashing_app.models import UserData
-from balance_cashing_app.views import _auto_create_balance
+from balance_caching_app.models import UserData
+from balance_caching_app.views import _auto_create_balance
 
 
 def update_balances():
     start_time = time.time()
-    # users is just for testing
+
     users = UserData.objects.all()
     for user in users:
-        time.sleep(300)
-        # user balance will come from portfolio async function
+        # time.sleep(300)
         user_id = user.user_identifier
         headers = {'Authorization': user_id}
         url = os.environ.get('AUTO_CACHING_URL')
-        response = requests.get(url+'api/assets', headers=headers)
+        response = requests.get(url + 'api/assets', headers=headers)
         total_balance = response.json()['total']
         data = {
             'id': user.user_identifier,
@@ -30,7 +29,6 @@ def update_balances():
 
     end_time = time.time()
     measured_time = end_time - start_time
-    print(measured_time)
     return measured_time
