+++ server/balances_history_cacher/balance_caching_app/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from balance_cashing_app.views import GetBalances, CreateBalance
+from balance_caching_app.views import GetBalances, CreateBalance
 
 urlpatterns = (
     path('', GetBalances.as_view()),
