+++ server/balances_history_cacher/balances_history_cacher/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for balances_history_casher project.
+ASGI config for balances_history_cacher project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'balances_history_casher.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'balances_history_cacher.settings')
 
 application = get_asgi_application()
