+++ server/balances_history_cacher/balances_history_cacher/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for balances_history_casher project.
+Django settings for balances_history_cacher project.
 
 Generated by 'django-admin startproject' using Django 3.2.9.
 
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
     'encrypted_model_fields',
-    'balance_cashing_app',
+    'balance_caching_app',
 
 ]
 
@@ -65,8 +65,7 @@ CORS_ALLOW_METHODS = [
     "PUT",
 ]
 
-
-ROOT_URLCONF = 'balances_history_casher.urls'
+ROOT_URLCONF = 'balances_history_cacher.urls'
 
 TEMPLATES = [
     {
@@ -85,7 +84,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'balances_history_casher.wsgi.application'
+WSGI_APPLICATION = 'balances_history_cacher.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/3.2/ref/settings/#databases
@@ -97,16 +96,14 @@ WSGI_APPLICATION = 'balances_history_casher.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'djongo',
-        'NAME': os.environ.get('DB_NAME'),
+        'NAME': 'balance_cache',
         'CLIENT': {
             'host': os.environ.get('DB_HOST_BALANCES_CACHING'),
-            'username': os.environ.get('DB_USERNAME_BALANCES_CACHING'),
             'password': os.environ.get('DB_PASSWORD_BALANCES_CACHING'),
             'authMechanism': 'SCRAM-SHA-1'
         },
     }
 }
-
 
 # Password validation
 # https://docs.djangoproject.com/en/3.2/ref/settings/#auth-password-validators
