+++ server/balances_history_cacher/balances_history_cacher/urls.py
@@ -1,4 +1,4 @@
-"""balances_history_casher URL Configuration
+"""balances_history_cacher URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
@@ -18,5 +18,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('balances/', include('balance_cashing_app.urls'))
+    path('balances/', include('balance_caching_app.urls'))
 ]
