+++ server/balances_history_cacher/balances_history_cacher/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for balances_history_casher project.
+WSGI config for balances_history_cacher project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'balances_history_casher.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'balances_history_cacher.settings')
 
 application = get_wsgi_application()
