+++ server/balances_history_cacher/requirements.txt
@@ -7,4 +7,4 @@ djongo==1.3.6
 dnspython==2.1.0
 pip-chill==1.0.1
 requests==2.26.0
-pymongo==3.12.0
+pymongo==3.12.1
